program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -13,7 +13,7 @@
 
 int main(void)
 {
-	int cur = 2, prv = 1, i, temp;
+	long int cur = 2, prv = 1, i, temp;
 
 	printf("1");
 	for (i = 1 ; i < 50 ; i++)
